source/requirements.txt
boto3
awscli
SQLAlchemy==1.3.15 # https://github.com/apache/airflow/issues/8211 # remove this if upgrading airflow from 1.10.9
psycopg2-binary
pandas==0.25.3
